# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/settings/profile": {
      "patch": {
        "operationId": "SettingsupdateProfile",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseSchema"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/change-password": {
      "post": {
        "operationId": "SettingschangePassword",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/set-password": {
      "post": {
        "operationId": "SettingssetPassword",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "operationId": "SettingsgetSettings",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AccountSettingsSchema"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SettingsupdateSettings",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AccountSettingsSchema"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      },
      "delete": {
        "operationId": "SettingsdeleteAccount",
        "parameters": [
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/feedback": {
      "post": {
        "operationId": "SettingssubmitFeedback",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization id (optional for user-scoped routes)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/report-problem": {
      "post": {
        "operationId": "SettingsreportProblem",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization id (optional for user-scoped routes)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Accept-Language",
            "description": "API message language (`en` or `sr`)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ],
              "default": "en"
            }
          },
          {
            "in": "header",
            "name": "x-lang",
            "description": "Alternative language header (same as Accept-Language)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "sr"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportProblemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Edge API",
    "description": "Digital signage CMS API. Success responses use `{ success: true, data }`. Errors use `{ success: false, error: { code, message, details? } }`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorEnvelopeSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorSchema"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "error",
          "path",
          "timestamp"
        ]
      },
      "UserResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "local",
              "google"
            ]
          },
          "hasPassword": {
            "type": "boolean"
          },
          "isSuperAdmin": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "provider",
          "hasPassword",
          "isSuperAdmin"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "phone": {
            "type": "string",
            "minLength": 1
          },
          "company": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "currentPassword",
          "password",
          "confirmPassword"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "password",
          "confirmPassword"
        ]
      },
      "AccountSettingsSchema": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "en",
              "sr"
            ]
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ]
          }
        },
        "required": [
          "language",
          "theme"
        ]
      },
      "UpdateAccountSettingsDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "en",
              "sr"
            ]
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ]
          }
        }
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "rating",
          "message"
        ]
      },
      "ReportProblemDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}